% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{indoorControlOptions}
\alias{indoorControlOptions}
\title{Options for an indoor floor control}
\usage{
indoorControlOptions(
  title = "Floor",
  descending = TRUE,
  max_height = 280,
  empty_label = "No levels"
)
}
\arguments{
\item{title}{Accessible title displayed above the floor buttons.}

\item{descending}{Display higher floors above lower floors.}

\item{max_height}{Maximum button-list height in CSS pixels.}

\item{empty_label}{Text displayed when the linked data set has no floors.}
}
\value{
An object of class \code{leaflet_indoor_control_options}.
}
\description{
Options for an indoor floor control
}
