% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-indoor-package.R
\docType{package}
\name{leaflet.indoor-package}
\alias{leaflet.indoor}
\alias{leaflet.indoor-package}
\title{leaflet.indoor: multi-level indoor maps with Leaflet}
\description{
The package adds map-local indoor data sets, accessible floor controls, and
Shiny proxy methods to \code{\link[leaflet:leaflet]{leaflet::leaflet()}} maps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aureliennicosiaulaval.github.io/leaflet-indoor/}
  \item \url{https://github.com/AurelienNicosiaULaval/leaflet-indoor}
  \item Report bugs at \url{https://github.com/AurelienNicosiaULaval/leaflet-indoor/issues}
}

}
\author{
\strong{Maintainer}: Aurélien Nicosia \email{116816597+AurelienNicosiaULaval@users.noreply.github.com}

Authors:
\itemize{
  \item Aurélien Nicosia \email{116816597+AurelienNicosiaULaval@users.noreply.github.com}
}

}
\keyword{internal}
