% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indoor.R
\name{removeIndoorControl}
\alias{removeIndoorControl}
\title{Remove an indoor floor control}
\usage{
removeIndoorControl(map, control_id = "indoor")
}
\arguments{
\item{map}{A Leaflet map or proxy.}

\item{control_id}{Control to remove.}
}
\value{
The input map or proxy.
}
\description{
Remove an indoor floor control
}
