% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indoor.R
\name{setIndoorLevel}
\alias{setIndoorLevel}
\title{Set the active indoor floor}
\usage{
setIndoorLevel(map, level, dataset_id = "indoor")
}
\arguments{
\item{map}{A Leaflet map or proxy.}

\item{level}{One non-missing floor value.}

\item{dataset_id}{Indoor data set to update.}
}
\value{
The input map or proxy.
}
\description{
Set the active indoor floor
}
